+++ app/src/main/java/com/example/bryan/whatsteddysname/activities/MainActivity.java
@@ -1,10 +1,12 @@
-package com.example.bryan.whatsteddysname;
+package com.example.bryan.whatsteddysname.activities;
 
 import android.content.Intent;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
+
+import com.example.bryan.whatsteddysname.R;
 
 public class MainActivity extends AppCompatActivity {
     private Button logInBtn;
