+++ app/src/main/java/com/example/bryan/whatsteddysname/asynctasks/ForgotPasswordContinuationTask.java
@@ -1,4 +1,4 @@
-package com.example.bryan.whatsteddysname;
+package com.example.bryan.whatsteddysname.asynctasks;
 
 import android.os.AsyncTask;
