+++ app/src/main/java/com/example/bryan/whatsteddysname/asynctasks/HttpGetRequest.java
@@ -1,4 +1,4 @@
-package com.example.bryan.whatsteddysname;
+package com.example.bryan.whatsteddysname.asynctasks;
 
 import android.app.ProgressDialog;
 import android.content.Context;
@@ -6,6 +6,8 @@ import android.content.Intent;
 import android.os.AsyncTask;
 import android.util.Log;
 import android.widget.Toast;
+
+import com.example.bryan.whatsteddysname.activities.SearchResultsActivity;
 
 import org.json.JSONException;
 import org.json.JSONObject;
