+++ app/src/main/res/layout/activity_collection.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.example.bryan.whatsteddysname.CollectionActivity"
+    tools:context="com.example.bryan.whatsteddysname.activities.CollectionActivity"
     android:focusableInTouchMode="true">
 
     <EditText
