+++ app/src/main/res/layout/activity_forgot_password.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/wtnbg"
-    tools:context="com.example.bryan.whatsteddysname.ForgotPasswordActivity"
+    tools:context="com.example.bryan.whatsteddysname.activities.ForgotPasswordActivity"
     android:focusableInTouchMode="true">
 
     <TextView
