+++ app/src/main/res/layout/activity_item.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.example.bryan.whatsteddysname.ItemActivity">
+    tools:context="com.example.bryan.whatsteddysname.activities.ItemActivity">
 
 
     <LinearLayout
