+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/wtnbg"
-    tools:context="com.example.bryan.whatsteddysname.MainActivity">
+    tools:context="com.example.bryan.whatsteddysname.activities.MainActivity">
 
     <TextView
         android:id="@+id/textView"
