+++ app/src/main/res/layout/activity_splash.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/wtnbg"
-    tools:context="com.example.bryan.whatsteddysname.SplashActivity">
+    tools:context="com.example.bryan.whatsteddysname.activities.SplashActivity">
 
     <TextView
         android:id="@+id/textView5"
